requirements.txt
# Web scraping dependencies
requests>=2.25.1
lxml>=4.9.0
cssselect>=1.2.0

# RSS feed generation
rfeed>=1.1.1
pytz>=2023.3

# Configuration handling
pyyaml>=5.4.1

# Date handling
python-dateutil>=2.8.2